# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "ekoabdulgoffar129@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:49/bpm-backend-laravel/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
...
        "/api/v1/skala_penilaian": {
            "get": {
                "tags": [
                    "SkalaPenilaian"
                ],
                "summary": "Get list of skala penilaian",
                "operationId": "afaef7ea1917cc916e51b4d8ce3e4066",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "skp_id": {
                                                "type": "integer"
                                            },
                                            "skp_skala": {
                                                "type": "integer"
                                            },
                                            "skp_deskripsi": {
                                                "type": "string"
                                            },
                                            "skp_tipe": {
                                                "type": "string"
                                            },
                                            "skp_status": {
                                                "type": "integer"
                                            },
                                            "skp_created_by": {
                                                "type": "string"
                                            },
                                            "skp_created_date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "skp_modif_by": {
                                                "type": "string"
                                            },
                                            "skp_modif_date": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SkalaPenilaian"
                ],
                "summary": "Create new skala penilaian",
                "operationId": "d01179738548aba8bbf6f296c464c6a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "skp_skala": {
                                        "type": "integer"
                                    },
                                    "skp_deskripsi": {
                                        "type": "string"
                                    },
                                    "skp_tipe": {
                                        "type": "string"
                                    },
                                    "skp_status": {
                                        "type": "integer"
                                    },
                                    "skp_created_by": {
                                        "type": "string"
                                    },
                                    "skp_created_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "skp_modif_by": {
                                        "type": "string"
                                    },
                                    "skp_modif_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Skala Penilaian created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/skala_penilaian/{id}": {
            "get": {
                "tags": [
                    "SkalaPenilaian"
                ],
                "summary": "Get skala penilaian by ID",
                "operationId": "8634886f025e9716c1938d37da0e8c8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID skala penilaian",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "skp_id": {
                                            "type": "integer"
                                        },
                                        "skp_skala": {
                                            "type": "integer"
                                        },
                                        "skp_deskripsi": {
                                            "type": "string"
                                        },
                                        "skp_tipe": {
                                            "type": "string"
                                        },
                                        "skp_status": {
                                            "type": "integer"
                                        },
                                        "skp_created_by": {
                                            "type": "string"
                                        },
                                        "skp_created_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "skp_modif_by": {
                                            "type": "string"
                                        },
                                        "skp_modif_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Skala Penilaian not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "SkalaPenilaian"
                ],
                "summary": "Update skala penilaian",
                "operationId": "28d240a851af174633fbce63a33fb595",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID skala penilaian",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "skp_skala": {
                                        "type": "integer"
                                    },
                                    "skp_deskripsi": {
                                        "type": "string"
                                    },
                                    "skp_tipe": {
                                        "type": "string"
                                    },
                                    "skp_status": {
                                        "type": "integer"
                                    },
                                    "skp_created_by": {
                                        "type": "string"
                                    },
                                    "skp_created_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "skp_modif_by": {
                                        "type": "string"
                                    },
                                    "skp_modif_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skala Penilaian updated successfully"
                    },
                    "404": {
                        "description": "Skala Penilaian not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SkalaPenilaian"
                ],
                "summary": "Delete skala penilaian",
                "operationId": "826e5598d7b31ec73b51091f4afcf795",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID skala penilaian",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skala Penilaian deleted successfully"
                    },
                    "404": {
                        "description": "Skala Penilaian not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    },
...
}